requirements.txt
aiomysql==0.1.1
aniso8601==9.0.1
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==39.0.2
databases==0.7.0
decorator==5.1.1
exceptiongroup==1.1.1
fastapi==0.95.0
flasgger==0.9.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
incremental==22.10.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
mysql-client==0.0.1
mysql-connector-python==8.0.32
outcome==1.2.0
packaging==23.0
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.2
pyrsistent==0.19.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
tornado==6.2
trio==0.22.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
zipp==3.15.0
zope.interface==6.0
